[IMP] currency_rate_live: Make it possible to launch tests more than 1 time
Never stop evolving.

FYI, the banxico tests currently work only 1 time when installing the module
with base is a single time. After the tests a CoA is automatically installed,
generating account.move.line.

When trying to launch the tests a second time, we try to change the currency of
the company to base.EUR or base.MXN, which is forbidden as move lines are already
generated for this company.

closes odoo/enterprise#2650