[REF] *: use 'frontend' system in non-website apps
* hr_contract_salary, sale_subscription, website_helpdesk

This commit does probably not do what is stated for all non-website apps
but it is a first step. It also uses the system in apps which could have
already used it but did not.

Part of https://github.com/odoo/enterprise/pull/3603
See https://github.com/odoo/odoo/pull/29442
task-1932066