[IMP] mrp_workorder: Work Order Form View
Small modifications in the form view of WO.
We add a new field (final_lot_id) on quality.check model
to track the finished product lot on workorder.

Previously, the field lot_id was filled with the component
lot in case of 'register_component_materials' and with the
finished product in case of other type. With this commit,
finished product lot and component lot have their own field

TaskID: 1903181

closes odoo/enterprise#3536