[IMP/REF] accounting-pocalypse yeaaahh
This commit adapts enterprise to the changes brought by odoo/odoo#33797.

Especially the merge of the following models
* account.invoice and account.move
* account.invoice.line and account.move.line
* account.voucher and account.move
* account.voucher.line and account.move.line

It was the opportunity for a big cleanup of the code, so it also restructures the whole account module, its different models/fields, the tests etc. for a better world and a better code readability.

==== Details of the changes ====
Please refer to the related commit in the community branch for the rationale behind this decision and the full log of all the changes it implied.

Here, are listed the changes that affect enterprise only.
* account.move: _set_invoice_lines method renamed into _get_invoice_lines
* account.move: _get_usable_mandate() method renamed into _sdd_get_usable_mandate()
* account.move: pay_with_mandate() method renamed into _sdd_pay_with_mandate()

Was task: 1917430

closes odoo/enterprise#4474