[ADD] industry_fsm_report: reports on tasks
Add new module allowing the creation of intervention reports for the
Field Service Management App. The report templates are created with
Studio.
It will be possible to assign a report template to be filled after  each task. The
worker has to fill the fields added via Studio on the template.

Task-1928442 and Task-1929681 Closes #3639