[FIX] account_online_sync: float_is_zero without journal currency
Before this commit, the online_sync_bank_statement wasn't working
correctly. This occurs when the currency was not defined in the journal
(currency is optional in journal). Indeed, when you call the
float_is_zero function with the currency's precision (0 as it's not
defined) an error is raised.

Now we take into account if currency is defined.

closes odoo/enterprise#3619