[FIX] l10n_mx_reports: fix test l10n installation
setting a mexican VAT number for the main company was raising an error in the test installing all localization, because l10n_mx_reports was installed after l10n_in that is adding a constraint validating the TIN of any partner located in India. Because of the odrer modules were installed, the country of the main partner was set to India, then the VAT validated, then when l10n_mx_reports was installed it was trying to change the VAT... raising the error from l10n_in.

Forcing the country in l10n_mx_reports seems a decent workaround for this particular case

closes odoo/enterprise#3638