[FIX] sign: make signature activity handling consistent
PR #3161 generates automatic activities for signature officers, but in
an inconsistent manner: the activity is created only for users with the
Sign Officer group, while the activity completion is generated for all
users (even portal users).

Additionally, the logic for the activity completion did not take into
account the fact that a partner could be linked to multiple users, which
would crash the activity feedback (expected singleton)

This patch fixes both issues.

Related to task 1912199.

closes odoo/enterprise#3653