[IMP] documents, sign: adds the direct sign action
This commit adds the option to use the direct sign feature
from the document's workflow actions.

Since this feature is only available for one record at a time,
this commit also makes it possible to define is a rule is limited
for a single record with the `on_single_record` computed field.

Task: #1929589

closes odoo/enterprise#3928

Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>