[FIX] voip: fallback on keypad when no phonecall found
Since last refactoring, users could only use the searchbar to filter
existing phonecalls and not to dial a phonenumber anymore. This was done
on purpose to trigger specific events for every tab.

This commits allows a fallback behaviour; when the search term does not
return any result, the user can hit the 'call' button and will be
redirected to the keypad where he can dial any number.

closes odoo/enterprise#3677