[IMP] web_mobile: close dialogs with Android's back-button
On mobile dialogs (modal or not) don't react correctly to the Android
back button: navigation happens but the dialog doesn't close.

This commit takes advantage of the custom event that the Android mobile
app provides when the user press the Android back-button (or perform the
related gesture).

Mobile Dialogs (modal or not) can then be closed by
pressing the back-button, without altering the browser history.

To enable the the BackButtonEventMixin it has to implement
on_attach_callback & on_detach_callback. The last is bound to the
Bootstrap event 'modal.bs.hidden' as we don't control ourself the moment
when the Dialog widget is detached from the DOM.

Taks ID: 1941120