[IMP] l10n_de_reports: added field Datev Identifier on the partner for Datev export
When migrating to Odoo, people want to keep their
old Datev identifiers.  As the identifier was set
to the id of the partner in Odoo, this was not possible.

So, we decided to add a field for the datev identifier
on the partner which allows to import the old number.
If it is not set we set it by default to 100000000 + the id.
This way, we also avoid conflicts with the account numbers.

You can also change the 100000000 by setting a system parameter
l10n_de.datev_start_count

Task ID : 1942108

closes odoo/enterprise#5083

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>