[IMP] mail, test_mail, *: add new email.cc.mixin on main models *: crm, project, maintenance, hr_recruitment
When a record is created from an email, cc can be lost. This commit
proposes a new mixin to keep cc on the record and allows to create
partner for each of them when sending a message from the chatter.

The mixin is added on the mail recors that can be created from mail.alias:

document
helpdesk.ticket
mrp.eco
quality.alert
hr.applicant
crm.lead
project.task
mail.channel
maintenance.equipment

Task: 1925001

closes odoo/enterprise#3689

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>