[ADD] hr_contract_sign: Request a signature from contract
Hr people often need to send some signature requests from a contract

For instance: commission on target, credit time, ...

There is a new wizard to select a hr responsible and a template to
sign. After validation, a signature request is send.

The signed documents are available on the private employee form,
for HR and managers, and on the 'My Profile' form, for employees

closes odoo/enterprise#3718

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>