[FIX] hr_contract_salary: Count all month for meal voucher
since commit 0557a2a090a51933636dbd6ebf2abc73f207321a the amount of
worked days is computed correctly according to the contract start.
i.e. Don't count the first 15 days in the month if your contract
started the 16th.

However when simulating the salary in the configurator, the temporary
contract's start date was set to the default (today).
That means that if you visit the configurator the 15th, only half of
the month is counted.

This is a problem for meal voucher where you receive a fixed amount
for each day. We want to display the amount for the all month and not
a different amount based on the day you visit the configurator.

Fix: set the simulation contract start date to the begining of the
month

closes odoo/enterprise#3734

Signed-off-by: "Yannick Tivisse (yti)" <yti@odoo.com>