[IMP] web_studio: improve field creation usability
This commit improves the overall usability of field creation using studio.
Main improvements are :

- When dropping a new field, that field will be selected automatically and
the properties tab will be opened right away for that field, instead of
having to click the field. Also, when a node is selected / clicked, label
of the node / field will be selected by default to ease the edition.
- When dropping a new one2many field, if there is no related many2one field
available, a warning will be shown instead of the empty drop-down.
- When dropping a new one2many field, if there are related many2one fields to
be displayed in drop-down, now model description will be displayed in the
brackets instead of technical names.

task-1917694

closes odoo/enterprise#4326

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>