[FIX] web_gantt: stop dragging when editing pill
Before this rev. a weird behaviour was happening when clicking on a pill to edit
it, then closing the opened dialog with ESC without moving the mouse. In this
use case, the pill was moving even though the mouse left click was not pressed.

This is due to the fact that, because the mouse hasn't moved, the drag is not
cancelled (by returning `false` in `drag` option of `draggable`). In this case,
the `mouseup` event, which had been manually stopped is used by jQuery to stop
the drag. Note that this code was used before the handling on `drag` just
described and is not useful anymore.

Task 1945936