[FIX] web_gantt: avoid crash when hovering resize badge
You have to be quick to reproduce this one: when dragging a pill, a resize badge
appears show you for how much time the pill will be moved. If you hover this
badge before it disappears, a traceback occurs.

This is due to the fact that the badge is in the cell (DOM speaking) but with
absolute position. As we use `elementsFromPoints` to display the + and the loupe
buttons, there are no cells hovered, hence the traceback.

Note that this commit cannot be tested because it uses elementsFromPoints which
is not correctly handled with QUnit tests.

Task 1945936

closes odoo/enterprise#3754

Signed-off-by: "David Monjoie (dmo)" <dmo@odoo.com>