[IMP] web_grid: supports readonly cell
Purpose of this commit is to be able to make some grid cell in a readonly
mode. The user will be able to see the details of the agreggated cell (list
view) but not edit the content. To do so, a new field type is introduced on
grid view architecture: `type="readonly"`. Defining such a field should be
done on a field having an agreagate psql function returning a boolean (like
`bool_and` and `bool_or` on boolean column).

The most known use case is the timesheet validation; we might want to freeze
the validated cell of timesheet grid view to avoid user to try to modify them.

Limitation:
For only boolean field are supproted to set a cell as readonly, as the ORM
does not support custom psql function (e.i.: a function computing is at least
one value of the agregate is higher that 4). Such a method can not be defined
on Odoo fields.

If we need more complex use case (customize color like list view, support more
field types, ...), futher dev should be done by framework team.

Task-32612

Co-authored-by: Mahendra Barad <mba@odoo.com>