[IMP] timesheet_grid: make validated grid cell readonly
When using timesheet validation, we want to prevent validated timesheet
edition. To do so, it is required to provide a aggregate function to
`validated` field of timesheet. We choose to freeze the cell when all the
timesheet in the aggregated cell are validated, so we used `bool_and`.
The `validated` field is added as `readonly` field of grid view.

To stay consistent on other views, we make them readonly too.

Task-32612

closes odoo/enterprise#1177

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>