[IMP] account: add automatic periodical VAT return
Add the possibility to automatically create the move that closes the VAT every 1-3 months. Take into accounts the different tax receivable/payable/advance account defined on the tax_group.
The created move will also contains a post message with the pdf and xml declaration that need to be sent to the government (for belgium)

closes odoo/enterprise#4076

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>