[FIX] web_studio: correctly update edited action
Since the action manager refactoring (odoo/enterprise@53437c7), the edited action was not
correctly updated.

To reproduce:
 - open Studio
 - edit the action (add a view type for example)
 - return in the action editor

The new view type will still be set as inactive.

This is because the action manager was still keeping a reference to the old action after
it has been updated. The reference is now correctly updated.

closes odoo/enterprise#3782

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>