[FIX] documents: makes share mail use owner id
Before this commit, uploading by mail wasn't using the
`owner_id` defined in the `documents.share` to create
the new document.

This commit makes so the new documents use the `share.owner_id`
if defined and uses the `share.create_uid` otherwise.

Additionally, removes the condition for `folder_id` as it is
a required field on `documents.share`.

closes odoo/enterprise#3791

Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>