[FIX] account_reports: show analytic menu
In 12.2 8cb9de9e6 the code of financial report has some improvements but
the analytic filter menu no longer appear.

If we have number of conditions (have a "analytic=True" report and be in
group analytic.group_analytic_accounting or in group
analytic.group_analytic_tags) we have `filter_analytic` set to True.

Before 8cb9de9e6 this caused `analytic` options to be set to true, now
did not.

opw-1949118
closes #3790

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>