[ADD] industry_fsm_stock: bridge module
This commit adds a bridge module between FSM and stock
modules. The goal is to automatically confirmed the stock
moves generated on Sales Order confirmation (SO coming from
FSM task). Indeed, we consider the stockable material as
already out of the stock when doing the SO. Rarely people
will handle a location "truck for work", so we don't
really supports that use case.

Task-1943792
Closes #3850