[IMP] appraisal: Remove dependency to survey + Misc improvements
1/ Remove dependency to survey.

This was badly integrated and most of the time people are requested to fill
a google form or something like that.

2/ Add default email templates to be sent on appraisal on the company

BTW, this is that way still possible to include a link to the survey
on the default email template.

3/ Add a reminder mechanism to warn employees, managers, ...
that an appraisal should be planned according to some configurations.

4/ Add a button on the employee form view to allow the manager
to request an appraisal.

5/ Add a button on the res.users 'My Profile' view to allow
an employee to request an appraisal.

closes odoo/enterprise#3854

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>