[FIX] helpdesk_crm: make previous_assigned_user determinist
In some case, test_team_assignation_randomly could break when
run localy (looks like it never breaks on runbot).
This is because the creation date of each helpdesk ticket is
exactly the same in this case.

It looks like this problem is recurent, sorting on a column
with identical values will deterministically give different
results beased on configuration (psql version?, config?)

This fix simply sort on ids too to avoid this problem.

closes odoo/enterprise#3853

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>