[IMP] l10n_be_hr_payroll: Add meal voucher report
Purpose
=======

Add a nice report for meal vouchers.
Allow to export a clean xls with meal voucher by employee and by month.

Specification
=============

A meal voucher should be granted for each normal work day of at least
4 hours. Work days are based on work entries.

Pivot view:
- rows: employees
- columns: number of meal vouchers by month
Default filter: this year.

An sql view report is added to compute meal voucher for each day.
It is used both in the pivot view and in the meal voucher
salary rule (through a computed field) as the quantity.

Note: In the pivot view, we would like months to be sorted
the more recent on the left. However this is currently not
supported neither by the pivot view, nor by `read_group` through
the `_order`.

TaskID: 1951586

closes odoo/enterprise#3857

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>