[IMP] sale_timesheet_enterprise: pause on timer
Impacted modules: sale_timesheet_enterprise, industry_fsm.

To improve timesheeting of tasks in Field Service and Project, this
commit add a "pause" button to their respective task form views.

This is reflected in the kanban view of tasks in Project, where a play
or pause icon is displayed accordingly to the corresponding task timer
state.

In order to simplify a redundant use of the same three fields in those
different buttons display conditions, we grouped them in a new
computed field, 'allow_timesheet_timer'.

Task-1943792