[IMP] l10n_be_hr_payroll: Add fiscal voluntarism on contracts
Purpose
=======

Sometimes you want to pay more taxes every month instead of paying
a larger amount after.
e.g. Withholding tax in PFI contract is very low every month, but you'll
have to pay more at the end of the year.

Specification
=============

Allow to voluntarily increase the withholding taxes.
Add a field "Fiscal Voluntary Rate" on the contract and if set,
use this rate instead of the basic amount ONLY if the amount with the new rate is higher.