[FIX] documents: open preview while editing
Forward port of cc58825 done in 66d3dabe had to disable a broken test in
order to merge. This fix correct the previous commit so the test is
restored.

Behavior change. In v12.0 the preview was put on hold until all saves
complete. In v12.1 the preview is canceled if there are saves.

closes odoo/enterprise#3871

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>