[IMP] hr_payroll: Add unpaid work engine
Purpose
=======

Currently, all work entries are considered paid when payslips
are computed.
However, some work entry types may not be paid (e.g. unpaid leave,
long sick leave).

Some work entries are paid for some salary structures but not in others.
e.g. In Belgium, "Breastfeeding Break" is not paid in the monthly
payslip, but counts for "13th month" and "double holiday".

There should be a way to choose which work entries (based on the type)
are paid/unpaid based on the payslip structure.

Specification
=============

Add a m2m field `unpaid_work_entry_type_ids` on payslip
structures.

When a payslip is computed, the amount corresponding to
unpaid work entries is deducted.
The deducted amount equals the number of unpaid days
multiplied by the average daily salary.

The average daily salary is computed as the average
wage over 3 months (the number of days in three months
is always constant = 13 weeks)