[IMP] web_mobile: add BackButtonEventMixin
Android devices have a dedicated back button (physical or virtual) but
sadly there is no standard event to target it.

Our mobile application provides an optional 'backbutton' event that can
be enabled in the webview, allowing us to customize its behavior (the
default being to go back in the history stack).

This commit introduces a BackButtonEventMixin that can be included in a
Widget to attach a custom event handler to the back-button. Its
implementation uses a stack to ensure that only the current widget will
react to the 'backbutton' event.

Remark: this mixin requires that the widget including it implements
correctly the `on_attach_callback` and `on_detach_callback` methods.

Taks ID: 1941120