[IMP] mrp_workorder: suggest lot on workorder
On a manufactoring order, 'checking availability' will make the reservation of
the components to consume. If those components are tracked by lot or serial numbers, the
reservation will also set a available lot/sn on the move lines and workorder lines.

This commit uses those lot numbers and quantity transfers on the
workorder lines to suggest them on the workorders tablet view.

To achieve the suggestion, quality checks are created with a linked workorder lines
The lot and quantity done of the checks are so the same as the lines ones.

In case of requesting the reservation after planning the production (applying
reservation on already created workorder lines and checks), we deleted all lines
not yet consumed. They are then recreated with the reservation.

The user can obiously still use the lot/sn he wants by changing it onto the
workorder.

Task : 1891864