[ADD] l10n_mx_edi: Add functionality from l10n_mx_edi_cancellation before
Why is this necessary?
In 2019 the SAT improved the Cancel process for CFDIs, and has some
restrictions to allow to cancel a document, e.g.:

- The CFDI does not have a refund related
- The CFDI does not have a payment related
- The amount total in the invoice is less that 5,000 MXN
- The customer must approve the cancellation

Here more information:
https://docs.google.com/presentation/d/1EuiDEx94xK7tVGGzo2WUGoLbAkvh6-M2sNAAB2zDPL4/edit?usp=sharing

Then, in some cases an invoice cannot be cancelled in the SAT, but in
the current process the invoice is cancelled in Odoo, and this is not
correct.

Now the invoice in Odoo must be cancelled only if the CFDI was cancelled
in the SAT, to make consistent the CFDI status with Odoo.

How was fixed?
The cancel button in Odoo is not available when the CFDI is valid in the
SAT system, now to cancel the CFDI must be press the button ``Request
Cancellation``, that will to try cancel the CFDI in the SAT system, and
if it could be cancelled, the invoice in Odoo also is cancelled.

If the invoice must wait for the customer approval, now was added a
cron, that is executed to verify which invoices are cancelled in the SAT
but not in Odoo, and in this case the cancel method from Odoo is called.

Also, was added a cron, that allows revert an invoice cancelled in Odoo,
if it is valid in the SAT, but not in Odoo.

Extra considerations:
Now finkok sends a notification, where said that when a CFDI is
cancelled, they do not ensure that it was cancel, only check if it could
be cancelled, and send the notification to the SAT, and after, the user
must ensure that the document was cancelled. This is supported now.

see also odoo/enterprise#4402