[FIX] purchase_mrp_workorder_quality: demo data
The quantity done on raw components is 0 instead of 1 for table head and
4 for table's legs

The quantity done for workorder line linked to a quality point or with a
tracked product are set to zero during the creation. Quantity is set through the
linked quality check once the user validate it. However the demo
data directly write on the workorder line the lot but not the quantity
done, it also not use the _next method that write the quantity done from
the check to the workorder line.

This commit write the quantity done on the workorder line that is write
on the move line during record production.