[IMP] l10n_be_hr_payroll: Add private car transport mode
Purpose
=======

Currently when an employee comes to work with his private car,
the reimbursed amount is provided with the `transport_mode_others`.
Any amount can be entered without restriction.
However CP200 has a rule to compute the reimbursed amount in
this case, based on the number of kilometers between home and the
work place.

Specification
=============

Add a transport mode "Uses private car".
Compute the amount to reimbursed based on `km_home_work`.
See task 1913545 for the computation rule.

Add a salary rule for this new reimbursed amount.

Hide "Other Reimbursed amount" on the contract form.