[IMP] hr_contract_salary: Ask first and last name
Purpose
=======
In the salary configurator people sometimes
enter "<Last name> <First name>"
and sometimes "<First name> <Last name>".

This makes the employee's names formatting heterogeneous.

We want to always have "<Last name> <First name>".

Specification
=============

In the salary configurator font end, ask the last
and first names separately than concatenate
them for backend.

For existing employees, we don't know how the name
is formatted, so keep the name in a single input.
(and it's impossible to have a clever heuristic)

closes odoo/enterprise#3945

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>