[FIX] web_mobile: inline elements shouldn't take all width
Before this commit, "Remind Before" field* wasn't properly aligned
because we forced oe_inline elements to take all width.

This rule was added to try to fix "Ordered Quantity" field in sale
order lines who doesn't fit the page because of the optionnal
unit of measure.
Anyway, this rule is now deleted because this issue shouldn't
be fixed in mobile only as the same issue is present on desktop.

(*) Calendar -> new event -> Edit -> Reminders -> Create and Edit

Task ID: 1944884

closes odoo/enterprise#3962

Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>