[IMP] mrp_workorder: multiple finished product lots
When registering lot or serial number for a component in a workorder,
the manufacturer can create it on on the fly. We would like to give this
ability a restriction depending on the picking type.
This commit blocks the creating by default but leave the possibility to
tick a new checkbox on the manufacturing picking type to allow the
creation of lot and serial number for components inside the production process.

Task : 1891864

closes odoo/enterprise#2909

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>