[FIX] web_enterprise: display navbar icon
Rev. 2324786 defaulted the configuration of the webclient as 'not
being in the home menu'. As a consequence, part of the code
updating the UI when switching from or to the home menu wasn't
executed anymore at webclient startup. As a consequence, the top
left icon in the navbar wasn't displayed when the webclient was
started on an action (not in the home menu).

As we now assume that we are not in the home menu at startup, we
directly use the correct icon in the menu template.

closes odoo/enterprise#4044

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>