[IMP] quality, mrp_workorder: check, point, workoder link
The structure between quality point, check and workorder is neat.
The point create the check and the check define the workorder line
test type and component. Also sometime the check could be create without
point.
- Use always test_type_id on each_model, test_type is always a related
of test_type_id.techincal name.
- Since the workorder line component and test type always depends on the
linked quality check, use a related on the workorder instead of a
compute.
- When a check is created from a point, test_type and component are not
required and could be used from the point that created it.
- test_type_id should not be a related to the point since check can be
created without point_id.