[IMP] ocn_client: return ocn_token to save it in the mobile app
This commit modifies the return of register_device method to allow
to get the ocn_token in javascript. So now we can save it in the
mobile (native) app to have a unique relation between registered
user and ocn_token.

This is needed for the mobile app to have unique key for each user.
It's required to fix a bug when you received notifications for another
user on the same instance as the current user.
The notification opened the given record with the current user
instead of the one for whom this notification was intended.

Task ID : 1941089