[REF] mrp_workorder: by-products as a check
Create a new type of step in order to register by-products. Also create
by-products without point and tracked by serial number or lot at the end
of the operation where they are produced(specified on the BoM).

Technically, a new test_type 'register by-product' is created. It will
replace the current check with test_type_id 'register consumed material
and component_is_byproduct boolean.

It also modify _defaults_from_workorder_lines and pass the test_type as
argument. The purpose is to find a check in the correct one2many. It
also directly create a missing workorder line if it could not find a
suitable one. It avoid to call twice the method by check if a line could
be find, if not create one and call the method again.