[IMP] web_studio: inline edition of selection values
Before this rev., when editing an existing selection value, a modal was
opened with inputs for the value and label. While not in debug mode,
only the label editable. The modal thus felt a bit empty.

When not in debug, the label edition is now inline and there is no modal
anymore.

Task 1916583

closes odoo/enterprise#3388

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>


Co-authored-by: Suraj Shukla <suh@odoo.com>