[IMP] hr_payroll: Change payslip name when date
Currently, the payslip name is automatically (onchange)
set to "[Structure default payslip name] - [Employee name] - [Payslip month and year]"
when the structure is changed.
It is not updated when the payslip date is changed.

With this commit, when the date of a payslip is changed,
the payslip name is updated.

Also, use the `format_date` function from tools
instead of using babel directly.