[REF] mrp_mps
The purpose of this commit is to redo master production schedule
with the purpose to:
- Have a better integration with existing records. For demand forecast
and replenishment it adds a tools in order to show the actual demand and
the actual replenishment. Also running the replenishment will deduce the
value to replenish with the existing replenishment already in progress.
- Flexibility on already launched forecast. If the user modify the
demand forecast or the quantity to replenish after the procurement
creation, it will allow the user to relaunch the cell in order to
complete replenishment.
- Multiple warehouses management
- Allow to define a custom number of periods. It would allow use case
with long manufacturing process
- The rules lead time is added to manufacturing and suppliers delay. It
is usefull if the transfer are long between 2 warehouses as example.

Visual effect for the end user:
- Remove the period from the control panel because the user could modify
it without modification on the min and max replenishment for each
schedule, it would completely modify the future forecast and create a
lot of inconsistent documents. (Still available in configuration).
- Use the total width on the screen in order to shpow the information
more clearly on the screen
- Apply button on schedule with a BoM linked to other MPS is replaced
by a direct reload on impacted schedule.
- Hover on replenish button that highlight the forecast that will be
run. The replenish button in control panel is based on lead time, the
replenish button on row, will run the next forecast.
- Tooltips on each row in order to explain the purpose.

closes odoo/enterprise#4551

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>