[FIX] web_gantt: fix slot insertion with DST
When we switch to DST, two unusual things happen:
- One day is missing one hour (1pm then 3pm)
- One day has the same hour twice (2pm then 2pm)

Before this commit, pills were placed in wrong slots on those days
because of the missing/duplicated slots.

To fix this, I changed the way we insert pills into slots, which
is now the same way we match pills into intervals for aggregation.

Because the variable was named "pillsInThisSlot", I thought I could
merge the two and only compute it once. However, one of them put
pills into slots (the new one), while the other actually put pills
into intervals (the existing one), so I changed the name of some
existing variables to make it more obvious which are referring to
slots and which are referring to intervals.