[IMP] documents: inspector on mobile
Documents inspector is represented on desktop as a right panel but it
requires some adjustment on mobile.

This commit implements it as a modal screen on mobile.

When the document selection change a bottom button is displayed in the
kanban view, allowing the user to open the inspector. If the user click
on a kanban record the inspector is directly opened for this document.

Task ID: 1917277