[IMP] web_gantt: allow to disable plan button
Removing the plan button whithout setting edit=false on gantt view is
sometime desireable. For example when dealing with objects that are
constrained to have a start and end date, the list of plan-able objects
is always empty. We add a plan option on the gantt vew to do so.

related task #1929079