[REM] web_studio: remove 'Enable Stages' feature
Currently, we have a feature to enable stages using studio on a kanban view.
But it is still a mere group by, not a pipeline. It is misleading as users
expect to get something similar to a project or crm pipeline.

Due to technical limitations (see task 1923279), we are not able to define
such a 'pipeline grouping' with studio. So with this commit, we simply get
rid of this feature.

task-1959224

closes odoo/enterprise#4152

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>