[IMP] hr_payroll: Display real hours/days in worked days line
1/ In worked days lines, we need to have whole (or half) numbers for
days.

In work entry type, we can specify the rounding type: up, down,
nearest or no rounding. If we select a rounding type, we can specify
if we allow half days (for example 3.5 days).

On work entry type, we can choose if it's an extra attendance, for
example for extra hours. The amount and number of days in work day
line will be set to 0. It will not impact other lines of worked days
lines.

All amount will be computed based on hours.

2/ hr_holidays: Compute real hours

The number of hours in time off were computed based on a mean of
number of hours in a week. It can engender problem if an employee
doesn't work with the same number of hours every day.

Now the number of hours are computed based on employee's calendar
and so the real number of hours are used.

closes odoo/enterprise#4170

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>