[IMP] mrp_workorder: onchange quantity to consume
Changing the field qty_producing on a workorder triggers an onchange
that update the quantities to produce on workorder lines. It also will
call the component compute method to change the component remaining
quantity.

The quantity to consume (qty_done) was the last field to be
updated by qty_producing.

This patch also remap the quality check quantities with their respective
workorder line 'quantity to consume' each time the workorder lines are
updated. This will ensure the suggested quantities on the workorder are
always good.

task : 1959158

closes odoo/enterprise#4129

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>