[FIX] l10n_be_hr_payroll: Fix auto increase wage
Commit ec86223 changed the employer onss
to the 2019 value.
But 41a8e9b introduced the 2018 value elsewhere
in parallel.
As the two commits did no conflicted at rebase
it was not seen when merging and the 2018 value
was left unmodified.

The consequence of this wrong value is
seen when changing the wage of a contract:
set the wage, then it is overwritten by
an onchange to a higher value.
Sadly this behaviour is wrong.

closes odoo/enterprise#4172

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>