[FIX] currency_rate_live: get latest rate instead of the dated one
When fetching xe_com page, a date parameter is given based on today's
date. If the timezone from where the request is done is ahead from
the xe_com server timezone, the page maybe empty.

For example, the request is made on 2019-04-23 at 02:00 am CEST, the date
at the xe_com location is still 2019-04-22, leading to an empty page.

With this commit, the date parameter is omitted, that way, the latest
currency rates are fetched.

closes odoo/enterprise#4173

Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>