[FIX] mrp_workorder: duplicate product in bom lines
In a bom, using the same product in two or more bom lines is allowed
but not managed correctly. Only one quality check is created for the
same product with the quantity of the first bom line. This obiously
lead to an issue in strict consumption.

This commit changes the way quality checks are created.
Before: A quality check was created by product
After: A quality check is created by move

Task : 1973904