[IMP] account_reports: phone should be clickable on follow-up reports
Purpose of this task is, on the follow-up reports,
the phone and mobile numbers of the contact are indicated.
the user can now directly call those numbers by clicking on them
So I made the phone number clickable on follow-up reports.

here we need to check for the voip  so added name on voip and if voip
installed then do the direct calling from voip.

Task- 1965862

closes odoo/enterprise#4077

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>