[IMP] pos_iot: Integrate electronic payment terminal into the POS
This commit create a new type of device: 'payment'.

Add the possibility to use the payment terminal selected into
the pos config for each payment method.

Add in the pos the possibility of using an electronic payment terminal
linked to the IoT Box by using the longpoolling of the IoT