[ADD] hr_referral: Motivate employees to make cooptation
Each employee has a personal dashboard with informations about him
cooptations (informations : his level in application referral, current
amount of points, ...). On dashboard there are also buttons to go to
available job positions to share, referral in progress or hired and
to the shop to spend points. The last button is made to send a mail
with a link with all job positions.

We can share job positions on Facebook, Twitter, Linkedin and direct
link. There are link_trackers on each generated link to give referral
points to correct employee.

When someone apply to a job with a shared link, the employee that
shared the link is added to applicant. The employee receives points
and notifications when the applicant reach new stages.

When employees reach a certain amount of points, he improve his
level and he can exchange points to gifts in the shop.
When an applicant is hired, the employee can choose a new 'friend'
to add it on his dashboard.

TaskID: 1967977

closes odoo/enterprise#4809

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>


Co-authored-by: edi-odoo <edi@odoo.com>
Co-authored-by: jbm-odoo <jbm@odoo.com>
Co-authored-by: RomainLibert <rli@odoo.com>
Co-authored-by: Yannick Tivisse <yti@odoo.com>