[IMP] project forecast: forecast publication
The goal is to allow user without access to the database to be notified of what has been forecasted on their working time. We can either send a single forecast which will then state start and end times as well as resource hours and taks and project id. If the employee is linked to a user (and has permissions to use the project app) he also get a link to see it's forecast on the gantt view.

The field published allow to track forecast publication. And the override of the write method allows the field to always be up-to-date when some important information have changed.

related task #1929079